+++ metricflow/engine/metricflow_engine.py
@@ -36,7 +36,7 @@ from metricflow.execution.executor import SequentialPlanExecutor
 from metricflow.logging.formatting import indent_log_line
 from metricflow.model.semantic_model import SemanticModel
 from metricflow.model.semantics.linkable_element_properties import LinkableElementProperties
-from dbt_semantic_interfaces.object_utils import pformat_big_objects
+from dbt_semantic_interfaces.pretty_print import pformat_big_objects
 from metricflow.random_id import random_id
 from metricflow.plan_conversion.column_resolver import DefaultColumnAssociationResolver
 from metricflow.plan_conversion.dataflow_to_execution import DataflowToExecutionPlanConverter
